feat: 보유 중인 주식 / 주식 리스트 클릭 시 바텀시트 올라오는 기능 구현 (#59)
- 보유 중인 주식, 주식 리스트 클릭 시 바텀시트 올라오도록 통합 작업 진행

---------
